refactor: remove MNCSparsityEstimation and integrate functionality into MncSketch
