+++ test/CMakeLists.txt
@@ -76,7 +76,6 @@ set(TEST_SOURCES
         runtime/local/datastructures/CSRMatrixTest.cpp
         runtime/local/datastructures/DenseMatrixTest.cpp
         runtime/local/datastructures/MncSketchTest.cpp
-        runtime/local/datastructures/MNCSparsityEstimationTest.cpp
         runtime/local/datastructures/FrameTest.cpp
         runtime/local/datastructures/MatrixTest.cpp
         runtime/local/datastructures/TaskQueueTest.cpp
